+++ src/main/java/com/dao/OrderMapper.java
@@ -6,11 +6,11 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface OrderMapper {
-    long countByExample(OrderExample example);
+    Integer countByExample(OrderExample example);
 
     int deleteByExample(OrderExample example);
 
-    int deleteByPrimaryKey(Long orderId);
+    int deleteByPrimaryKey(Integer orderId);
 
     int insert(Order record);
 
@@ -18,7 +18,7 @@ public interface OrderMapper {
 
     List<Order> selectByExample(OrderExample example);
 
-    Order selectByPrimaryKey(Long orderId);
+    Order selectByPrimaryKey(Integer orderId);
 
     int updateByExampleSelective(@Param("record") Order record, @Param("example") OrderExample example);
