+++ src/main/java/com/service/MessageService.java
@@ -13,7 +13,7 @@ public interface MessageService {
      *1.摄影师根据自己的ID，获取到用户给自己的留言
      * @param photographerId
      */
-    public List<U_to_p_message> getUMessages(Long photographerId);
+    public List<U_to_p_message> getUMessages(Integer photographerId);
 
     /**
      *2.1用户给摄影师留言。
@@ -24,5 +24,5 @@ public interface MessageService {
      * 2.2用户根据自己的ID，获取到摄影师给自己的回复
      * @param
      */
-    public List<P_to_u_message> getPMessages(Long userId);
+    public List<P_to_u_message> getPMessages(Integer userId);
 }
